rfa: move methods to an utils file
